requirements.txt
asgiref==3.11.0
async-generator==1.10
attrs==25.4.0
certifi==2026.1.4
coverage==7.13.2
django==6.0.1
django-debug-toolbar==6.2.0
django-extensions==4.1
h11==0.16.0
idna==3.11
importlib-metadata==8.7.1
mypy-extensions==1.1.0
outcome==1.3.0.post0
packaging==26.0
pip==25.3
pysocks==1.7.1
selenium==4.40.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.5
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing-extensions==4.15.0
urllib3==2.6.3
websocket-client==1.9.0
wsproto==1.3.2
zipp==3.23.0
